requirements.txt
imapclient>=3.0
pyyaml>=6.0
youtube-transcript-api>=0.6
yt-dlp>=2024.1
pymupdf4llm>=0.0.10
